pyproject.toml
[tool.poetry]
name = "openai-image-mcp"
version = "0.2.0"
description = "OpenAI GPT-Image-1 MCP Server - Generate images using OpenAI's latest image model"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "openai_image_mcp", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.0.0"
mcp = {extras = ["cli"], version = "^1.0.0"}
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
httpx = "^0.27.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^23.0.0"
isort = "^5.0.0"
mypy = "^1.0.0"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
openai-image-mcp = "openai_image_mcp.server:main"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"